+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## In progress
 
+- Fix "reroll" button in Ironsworn oracle-roll chat cards
+
 ## 1.15.6
 
 - Cap action scores at 10 ([#371](https://github.com/ben/foundry-ironsworn/pull/371))